// dist/partie.d.ts
declare class Partie {
    titre: string;
    image: string;
    contenu: string;
    constructor(titre: string, image: string, contenu: string);
    getPartie(): {
        titre: string;
        image: string;
        contenu: string;
    };
}
export { Partie };
//# sourceMappingURL=partie.d.ts.map